Issue OS-157 DefinitionManager modified: added getAllSchemaDefinations(), getSchemaDefination(title)
